requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
cryptography==43.0.1
